pyproject.toml
[project]
name = "publicdataprotal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==6.0.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "apscheduler>=3.11.2",
    "attrs==26.1.0",
    "authlib>=1.6.9",
    "blinker==1.9.0",
    "cachetools==7.0.5",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.6",
    "click==8.3.1",
    "colorama==0.4.6",
    "cryptography==46.0.5",
    "fastapi==0.135.1",
    "folium>=0.20.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "google-api-core==2.30.0",
    "google-api-python-client==2.193.0",
    "google-auth==2.49.1",
    "google-auth-httplib2==0.3.0",
    "google-auth-oauthlib==1.3.0",
    "googleapis-common-protos==1.73.0",
    "h11==0.16.0",
    "httplib2==0.31.2",
    "idna==3.11",
    "jinja2==3.1.6",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "markupsafe==3.0.3",
    "narwhals==2.18.0",
    "numpy==2.4.3",
    "oauthlib==3.3.1",
    "packaging==26.0",
    "pandas==2.3.3",
    "pillow==12.1.1",
    "proto-plus==1.27.1",
    "protobuf==6.33.6",
    "pyarrow==23.0.1",
    "pyasn1==0.6.3",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings>=2.13.1",
    "pydeck==0.9.1",
    "pyparsing==3.3.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2026.1.post1",
    "referencing==0.37.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rpds-py==0.30.0",
    "six==1.17.0",
    "smmap==5.0.3",
    "starlette==0.52.1",
    "streamlit==1.55.0",
    "streamlit-folium>=0.27.1",
    "tenacity==9.1.4",
    "toml==0.10.2",
    "tornado==6.5.5",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
    "uvicorn==0.42.0",
    "watchdog==6.0.0",
]
